[SPARK-51502][SQL][TESTS] Move collations test to collations package
### What changes were proposed in this pull request?
Move collations test into collations package where most collation test suites already are located.

### Why are the changes needed?
For consistency purposes, also not to bloat the sql folder with too many suites.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests should suffice.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #50268 from stefankandic/moveCollTests.

Authored-by: Stefan Kandic <stefan.kandic@databricks.com>
