+++ sql/core/src/test/scala/org/apache/spark/sql/collation/CollationExpressionWalkerSuite.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql
+package org.apache.spark.sql.collation
 
 import java.sql.Timestamp
 
@@ -23,6 +23,7 @@ import org.apache.spark.SparkFunSuite
 import org.apache.spark.sql.catalyst.analysis.ExpressionBuilder
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.expressions.variant.ParseJson
+import org.apache.spark.sql.classic
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.internal.types._
 import org.apache.spark.sql.test.SharedSparkSession
